% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGAM_Func.R
\name{ac_remove_fun}
\alias{ac_remove_fun}
\title{Name: ac_remove_fun}
\usage{
ac_remove_fun(data, cols, cors, cutoff)
}
\arguments{
\item{data}{Dataframe with species abundance values and time-series predictor columns.}

\item{cols}{Columns of data that will be modeled using the GAM.}

\item{cors}{Correlation structure class.}
}
\value{
List of whether significant AC was detected in the GAM residuals of each species.
}
\description{
Description: Fits a GAM to each species and then checks to see if significant AC is still prevalent in GAM residuals
}
