% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGAM_Func.R
\name{get_upper_tri}
\alias{get_upper_tri}
\title{# Name: get_upper_tri}
\usage{
get_upper_tri(mat)
}
\arguments{
\item{mat}{Matrix of p-values where the upper triangle contains adjusted p-values.}
}
\value{
: Matrix of p-values where the upper triangle contained adjusted p-values and the lower triangle contains NAs.
}
\description{
# Description: Keeps upper triangle of adjacency matrix and makes the lower triangle NAs
}
