% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGAM_Func.R
\name{netGAM.df}
\alias{netGAM.df}
\title{Name: netGAM.df}
\usage{
netGAM.df(df, MOY, MCount, clrt = TRUE)
}
\arguments{
\item{df}{Species abundance dataframe with samples as rows species as columns.}

\item{MOY}{Vector that specifies the month of year for each sample (row) in the dataframe (i.e. 1-12).}

\item{MCount}{Vector that specifies the day of the time-series for each sample (row) in the dataframe (e.g. 1-200).}

\item{clrt}{If TRUE, the clr transformation in the compositions package is used to clr transform the input species abundance dataframe prior to GAM transformation (default is TRUE). If FALSE, the species abundance data are not clr transformed prior to GAM transformation.}
}
\value{
GAM-transformed dataframe (rows = samples, columns = species)
}
\description{
Description: Fits a gamm to each species in a species abundance dataset and returns a dataframe of the residuals of each gamm. The returned dataframe can be thought of as a species abundance dataframe with a reduced influence of time on the species abundance values. This dataframe can be used as input in downstream network analyses.
}
