% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGAM_Func.R
\name{works_fun}
\alias{works_fun}
\title{Name: works_fun}
\usage{
works_fun(data, cols, cors)
}
\arguments{
\item{data}{Dataframe with species abundance values and time-series predictor columns.}

\item{cols}{Columns of data that will be modeled using the GAM.}

\item{cors}{Correlation structure class.}
}
\value{
T/F list of whether errors popped up during GAM fitting
}
\description{
Description: Checks to see if the GAM can be fit to each species abundance vector in the input dataframe
}
