% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGAM_Func.R
\name{netGAM.network}
\alias{netGAM.network}
\title{Name: netGAM.network}
\usage{
netGAM.network(df, MOY, MCount, clrt = TRUE, method = "glasso", pvalue = NULL)
}
\arguments{
\item{df}{Species abundance dataframe with samples as rows species as columns.}

\item{MOY}{Vector that specifies the month of year for each sample (row) in the dataframe (i.e. 1-12).}

\item{MCount}{Vector that specifies the day of the time-series for each sample (row) in the dataframe (e.g. 1-200).}

\item{clrt}{If TRUE, the clr transformation in the compositions package is used to clr transform the input species abundance dataframe prior to GAM transformation (default is TRUE). If FALSE, the species abundance data are not clr transformed prior to GAM transformation.}

\item{method}{Networking method to use (default is glasso). "glasso" = graphical lasso network constructed with the "batch.pulsar" function in the pulsar package with StARS selection; "scc" = spearman correlation network constructed with the "corr.test" function in the psych package; "pcc" = pearson correlation network constructed with the "corr.test" function in the psych package.}

\item{pvalue}{P-value cutoff for deciding whether or not an edge exists (default is NULL). P-values in corrleation networks are bonferroni-adjusted prior to declaring cutoff. P-value only needed for scc and pcc networks.}
}
\value{
Adjacency matrix of network predicitons (1 = edge, 0 = no edge)
}
\description{
Description: Fits a gamm to each species in a species abundance dataset, extracts the gamm residuals for each species, runs a network analysis on the gamm residuals, and returns an adjacency matrix of network-predicted associations.
}
